prow: migrate ProwJob structs to prow/apis/
