+++ prow/pjutil/pjutil.go
@@ -71,8 +71,10 @@ func NewProwJob(spec kube.ProwJobSpec, labels map[string]string) kube.ProwJob {
 	}
 
 	return kube.ProwJob{
-		APIVersion: "prow.k8s.io/v1",
-		Kind:       "ProwJob",
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "prow.k8s.io/v1",
+			Kind:       "ProwJob",
+		},
 		ObjectMeta: metav1.ObjectMeta{
 			Name:   uuid.NewV1().String(),
 			Labels: allLabels,
